% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opti.R
\name{angle}
\alias{angle}
\title{Angle between two vectors}
\usage{
angle(a, b)
}
\arguments{
\item{a}{first vector}

\item{b}{second vector}
}
\value{
angle between the two vectors in radiants
}
\description{
Angle between two vectors
}
