% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unconstrained_gallery.R
\name{function_gallery}
\alias{function_gallery}
\title{Unconstrained problems}
\usage{
function_gallery(exp, n)
}
\arguments{
\item{exp}{string that identifies the experiment in the gallery}

\item{n}{dimension of the problem}
}
\value{
a list with suggested starting point, objective and gradient functions
}
\description{
Gallery of unconstrained optimization problems from different sources. They are all based on sum of functions, where the number of terms is related to the size of the problem.
}
\details{
e.g.,
\tabular{lll}{
Name \tab Function \tab Reference \cr
1 \tab Strictly Convex 1 \tab Raydan 1997 \cr
2 \tab Strictly convex 2 \tab Raydan 1997 \cr
3 \tab Brown almost linear \tab Raydan 1997 \cr
4 \tab Trigonometric \tab Raydan 1997 \cr
5 \tab Broyden tridiagonal \tab Raydan 1997 \cr
7 \tab Extended Rosenbrock \tab Raydan 1997 \cr
8 \tab Penalty 1 \tab Raydan 1997 \cr
10 \tab Variably dimensioned \tab Raydan 1997 \cr
11 \tab Extended Powell \tab Raydan 1997 \cr
12 \tab Generalized Rosenbrock \tab Raydan 1997 \cr
hager \tab Hager \tab Andrei 2008 \cr
gentridiag1 \tab Generalized tridiagonal 1 \tab Andrei 2008 \cr
genWH \tab Generalized White and Holst \tab Andrei 2008 \cr
griewank \tab Griewank \tab website \cr
ext-freud-roth \tab Extended Freudenstein and Roth \tab Andrei 2008 \cr
ext-WH \tab Extended White and Holst \tab Andrei 2008 \cr
ext-beale \tab Extended Beale 5 \tab Andrei 2008 \cr
pert-quad \tab Perturbed quadratic \tab Andrei 2008 \cr
diag1 \tab Diagonal 1 \tab Andrei 2008 \cr
diag2 \tab Diagonal 2 \tab Andrei 2008 \cr
diag3 \tab Diagonal 3 \tab Andrei 2008 \cr
ext-tridiag1 \tab Extended tridiagonal 1 \tab Andrei 2008 \cr
ext-tet \tab Extended TET (three exponential terms) \tab Andrei 2008 \cr
gen-tridiag2 \tab Generalized tridiagonal 2 \tab Andrei 2008 \cr
diag4 \tab Diagonal 4 \tab Andrei 2008 \cr
ext-himmelblau \tab Extended Himmelblau \tab Andrei 2008 \cr
ext-psc1 \tab Extended PSC1 \tab Andrei 2008 \cr
fh1 \tab Full Hessian FH1 \tab Andrei 2008 \cr
fh2 \tab Full Hessian FH2 \tab Andrei 2008 \cr
}
}
\examples{
function_gallery("genWH", 10)
}
