% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opti.R
\name{gtbb}
\alias{gtbb}
\title{Global TBB}
\usage{
gtbb(
  x,
  obj,
  grad,
  p = gtbb_control(),
  quadratic = FALSE,
  reltol = TRUE,
  info = FALSE
)
}
\arguments{
\item{x}{starting vector}

\item{obj}{objective function. Must be of the form obj(x)}

\item{grad}{gradient function. Must be of the form grad(x)}

\item{p}{control parameters. See gtbb_control for more details}

\item{quadratic}{whether the objective function is quadratic or not, default = FALSE}

\item{reltol}{whether the tolerance is multiplied by the norm of the first gradient (TRUE) or not. default = TRUE}

\item{info}{save information of the whole method, default = FALSE}
}
\value{
minimizer and some information: number of function evaluations, number of iterations, last gradient norm, tolerance, final objective value.
It optionally returns the aforementioned information about all the iterations.
}
\description{
Gradient method with nonmonotone line search for continuously differentiable functions.
}
\details{
It is endowed with the TBB stepsize, a Barzilai--Borwein like stepsize with an additional parameter, called target.
}
\examples{
A <- c(1:100)
x <- rep(1:100)

obj <- function(x) sum(A * x^2)/2
grad <- function(x) A * x

gtbb(x, obj, grad, quadratic = TRUE)
}
