% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opti.R
\name{gtbb_control}
\alias{gtbb_control}
\title{Parameters of global TBB}
\usage{
gtbb_control(
  alpha_min = 1e-08,
  alpha_max = 1e+30,
  alpha = NULL,
  gamma = 1e-04,
  M = 5,
  tol = 1e-06,
  iter_max = 10000,
  iter_int_max = 30,
  target_option = "bb1",
  ...
)
}
\arguments{
\item{alpha_min}{min value for the stepsize, default = 1e-30}

\item{alpha_max}{max value for the stepsize, default = 1e30}

\item{alpha}{initial value for the stepsize, default = 1}

\item{gamma}{line search param, default = 1e-4}

\item{M}{number of stored previous function evaluations, default = 10}

\item{tol}{tolerance for the stopping criterion, default = 1e-6}

\item{iter_max}{maximum number of iterations, default = 1e4}

\item{iter_int_max}{maximum number of iterations for the backtracking, default = 100}

\item{target_option}{string or function. See target_gallery for more details; default = "bb1"}

\item{...}{additional arguments (not needed at the moment)}
}
\value{
list with parameters for the
}
\description{
Parameters of global TBB
}
\details{
Other stepsizes:
\tabular{ll}{
Target \tab Description \cr
bb1 \tab Barzilai -- Borwein 1 \cr
bb2 \tab Barzilai -- Borwein 2 \cr
abb \tab Adaptive BB with threshold 0.8 \cr
homo_bb \tab Homogeneous BB \cr
homo_abb \tab Adaptive homo BB with threshold 0.8 \cr
}
}
\examples{
gtbb_control()
}
