% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unconstrained_gallery.R
\name{function_gallery}
\alias{function_gallery}
\title{Unconstrained problems}
\usage{
function_gallery(exp, n)
}
\arguments{
\item{exp}{string that identifies the experiment in the gallery}

\item{n}{dimension of the problem}
}
\value{
a list with suggested starting point, objective and gradient functions
}
\description{
Gallery of unconstrained optimization problems from different sources. They are all based on sum of functions, where the number of terms is related to the size of the problem.
}
\details{
e.g.,
\tabular{lll}{
Name \tab Function \tab Reference \cr
2 \tab Strictly convex 2 \tab Raydan 1997\cr
}
}
\examples{
function_gallery("genWH", 10)
}
