% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opti.R
\name{target_gallery}
\alias{target_gallery}
\title{Gallery of targets}
\usage{
target_gallery(target_option)
}
\arguments{
\item{target_option}{string corresponding to one of the available targets (add which)}
}
\value{
function with three arguments: angle between $s_k$ and $y_k$, inverse BB2 $\|y_k\|^2/s_k^Ty_k$, number of iterations
}
\description{
Gallery of targets
}
\details{
e.g.,
\tabular{ll}{
Target \tab Description\cr
cotan1_1 \tab cotangent step with $q=r=1$ \cr
}
}
\examples{
target_gallery("cotan1_1")
}
